Raw Image UI 에 추가
Nrrd 파일 Application.persistentDataPath 경로에 저장
