+++ Task1/Assets/Scenes/SampleScene.unity
@@ -697,6 +697,7 @@ RectTransform:
   m_Children:
   - {fileID: 1322544998}
   - {fileID: 1401663193}
+  - {fileID: 641026458}
   - {fileID: 651002156}
   - {fileID: 836786730}
   m_Father: {fileID: 0}
@@ -724,7 +725,7 @@ MonoBehaviour:
   rowContent: {fileID: 1426277291}
   seriesContent: {fileID: 1498893611}
   searchContent: {fileID: 651002155}
-  seriesText: {fileID: 502153540}
+  seriesButton: {fileID: 1078709950}
   inputField: {fileID: 1055696613}
   searchText: {fileID: 1339072955}
   keyword: 
@@ -2343,7 +2344,7 @@ MonoBehaviour:
     m_PersistentCalls:
       m_Calls:
       - m_Target: {fileID: 514762595}
-        m_TargetAssemblyTypeName: DicomTableview, Assembly-CSharp
+        m_TargetAssemblyTypeName: DicomImageViewer, Assembly-CSharp
         m_MethodName: OnClickSearch
         m_Mode: 1
         m_Arguments:
@@ -2496,8 +2497,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -21.4836, y: -55.18643}
-  m_SizeDelta: {x: 766.624, y: 324.7148}
+  m_AnchoredPosition: {x: 237, y: -41.458}
+  m_SizeDelta: {x: 492.6343, y: 344.13}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1401663194
 MonoBehaviour:
@@ -2817,7 +2818,7 @@ RectTransform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 502153539}
+  - {fileID: 1078709953}
   m_Father: {fileID: 1282244585}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 1}
