--- Task1/Assets/Scripts/DicomImageViewer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f2c3c2fe4a1c75e4b82ddf1f0eff50b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
