--- Task1/Assets/VolumeData.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 746ac9c28c8e4dc49b6a34ead8f1518c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
